requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.3
psycopg2-binary==2.9.7
openpyxl==3.1.2
flask-cors==4.0.0
gunicorn==20.1.0
python-dotenv==1.0.0
# Werkzeug is managed by Flask; avoid pinning to prevent dependency conflicts.
